requirements.txt
aiohttp
websockets
nats-py
asyncpg
python-dotenv
pandas
numpy